--- ObjLoader/Loaders/IMaterialLibraryLoader.cs
@@ -1,9 +1,0 @@
-using System.IO;
-
-namespace ObjLoader.Loader.Loaders
-{
-    public interface IMaterialLibraryLoader
-    {
-        void Load(Stream lineStream);
-    }
-}
