+++ ObjLoader/Loaders/IObjLoaderFactory.cs
@@ -2,7 +2,6 @@ namespace ObjLoader.Loader.Loaders
 {
     public interface IObjLoaderFactory
     {
-        IObjLoader Create(IMaterialStreamProvider materialStreamProvider);
         IObjLoader Create();
     }
 }
