+++ ObjLoader/Loaders/MaterialLibraryLoaderFacade.cs
@@ -2,18 +2,16 @@ namespace ObjLoader.Loader.Loaders
 {
     public class MaterialLibraryLoaderFacade : IMaterialLibraryLoaderFacade
     {
-        private readonly IMaterialLibraryLoader _loader;
-        private readonly IMaterialStreamProvider _materialStreamProvider;
+        private readonly MaterialLibraryLoader _loader;
 
-        public MaterialLibraryLoaderFacade(IMaterialLibraryLoader loader, IMaterialStreamProvider materialStreamProvider)
+        public MaterialLibraryLoaderFacade(MaterialLibraryLoader loader)
         {
             _loader = loader;
-            _materialStreamProvider = materialStreamProvider;
         }
 
         public void Load(string materialFileName)
         {
-            using var stream = _materialStreamProvider.Open(materialFileName);
+            using var stream = _loader.Open(materialFileName);
 
             if (stream != null)
                 _loader.Load(stream);
